ADD: separate start file created for start server logic - created separate startServer function in start.js - moved start server code from server.js to start.js - moved testing db connection code from db.js to start.js - added db sync code inside startServer function in start.js - run startServer
